+++ src/Containers/Interfaces/IInventory.cs
@@ -81,15 +81,6 @@ namespace TheChest.Inventories.Containers.Interfaces
         bool TryAdd(params T[] items);
 
         /// <summary>
-        /// <para> Adds an item in a avaliable slot </para>
-        /// <para> This method return will change to void in future versions. </para>
-        /// <para> Use <see cref="CanAdd(T[])"/> to check if the item can be added before calling this method.</para>
-        /// </summary>
-        /// <param name="item">item to be added</param>
-        /// <returns><see langword="true"/> if the <paramref name="item"/> could be added</returns>
-        [Obsolete("Use Add(params T[]) instead. This method will be removed in future versions.")]
-        bool Add(T item);
-        /// <summary>
         /// Adds and array of items in a avaliable slot
         /// </summary>
         /// <param name="items">Array of items to be added to any avaliable slot found</param>